Refactored updater. Sequential secion and item animations working better.
