+++ IdentifierPath.swift
@@ -1,0 +1,83 @@
+//
+//  IdentifierPath.swift
+//  FOCollections
+//
+//  Created by Daniel Krofchick on 2017-06-06.
+//  Copyright © 2017 Figure1. All rights reserved.
+//
+
+import Foundation
+
+struct IdentifierPath {
+    
+    var identifiers: [String]
+    
+    init(identifiers: [String]) {
+        self.identifiers = identifiers
+    }
+    
+    init(item: String, section: String) {
+        self.identifiers = [section, item]
+    }
+    
+    public subscript(range: CountableClosedRange<Int>) -> IdentifierPath? {
+        if range.lowerBound < 0 || range.upperBound > identifiers.count || identifiers.count == 0 {
+            return nil
+        } else {
+            return IdentifierPath(identifiers: Array(identifiers[range]))
+        }
+    }
+    
+    public subscript(index: Int) -> IdentifierPath? {
+        return self[index...index]
+    }
+    
+    public subscript(range: CountableRange<Int>) -> IdentifierPath {
+        var result = [String]()
+        
+        identifiers.enumerated().forEach {
+            (index, identifier) in
+            if index >= range.lowerBound && index <= range.upperBound {
+                result.append(identifier)
+            }
+        }
+        
+        return IdentifierPath(identifiers: result)
+    }
+    
+}
+
+extension IdentifierPath: Hashable {
+    
+    var hashValue: Int {
+        get {
+            return identifiers.reduce("", {
+                (result, value) -> String in
+                return result + value
+            }).hashValue
+        }
+    }
+    
+}
+
+extension IdentifierPath: Equatable {
+    
+    public static func ==(lhs: IdentifierPath, rhs: IdentifierPath) -> Bool {
+        guard
+            lhs.identifiers.count == rhs.identifiers.count
+            else {
+                return false
+        }
+        
+        return lhs.identifiers == rhs.identifiers
+    }
+    
+}
+
+extension IdentifierPath: CustomStringConvertible {
+    
+    public var description: String {
+        return String(describing: identifiers)
+    }
+    
+}
