+++ Move.swift
@@ -1,0 +1,19 @@
+//
+//  Move.swift
+//  FOCollections
+//
+//  Created by Daniel Krofchick on 2017-06-06.
+//  Copyright © 2017 Figure1. All rights reserved.
+//
+
+struct Move {
+    
+    var from: StatePath
+    var to: StatePath
+    
+    init(from: StatePath, to: StatePath) {
+        self.from = from
+        self.to = to
+    }
+    
+}
